Added Matrix4f and Matrix4d and added tests for RotationMatrixf.

+++ EfficientRealtimeJava3D/src/Matrix3f.java
@@ -574,7 +574,7 @@ public class Matrix3f implements java.io.Serializable
 		}
 	}
 
-	public double determinant()
+	public float determinant()
 	{
 		return (m00 * m11 * m22 + m01 * m12 * m20 + m02 * m10 * m21 - m02 * m11
 				* m20 - m00 * m12 * m21 - m01 * m10 * m22);
